requirements.txt
fastapi
uvicorn
python-multipart
pyannote.audio==3.1.1
torch
torchaudio

